Fix package name of "flow" API
