+++ pkg/transformation/pipeline/pipeline.go
@@ -20,7 +20,7 @@ import (
 	"fmt"
 	"log"
 
-	"github.com/triggermesh/triggermesh/pkg/apis/transformation/v1alpha1"
+	"github.com/triggermesh/triggermesh/pkg/apis/flow/v1alpha1"
 	"github.com/triggermesh/triggermesh/pkg/transformation/pipeline/common/storage"
 	"github.com/triggermesh/triggermesh/pkg/transformation/pipeline/transformer"
 	"github.com/triggermesh/triggermesh/pkg/transformation/pipeline/transformer/add"
